+++ Vault/Sources/VaultiOS/Views/Detail/SecureNote/SecureNoteDetailView.swift
@@ -69,7 +69,6 @@ struct SecureNoteDetailView: View {
                 isShowingDeleteConfirmation: $isShowingDeleteConfirmation,
                 navigationPath: $navigationPath,
                 presentationMode: presentationMode,
-                persistedLockState: viewModel.editingModel.initialDetail.lockState,
             ) {
                 if viewModel.isInEditMode {
                     noteContentsEditingSection